design: strategy + factory (yaml configs).
Each different load generator now is a LoadGenerationStrategyI.
Instantiation uses a simple factory that takes the config file
and unmarshals it into the necessary load generator.

Added commandline options for specifying the type of load to be
generated.

Load generator config files are written in YAML and have fixed names
and are stored in the project's root directory.
Configuration files are located in configs/.

Code compilation and Jar creation is now performed using Gradle.
Currently, build.gradle is written such that a fat jar is created.
This is to make sure that all the dependencies are bundled together
with the application.

Updated gitignore to ignore .gradle, build/ and .idea/.

Updated Dockerfile to now just run the system load generator driver.
Added an entrypoint script that runs the driver code (gradle run)
and appends the commandline arguments.

Added settings.gradle file to include the name of the project.
Without this, gradle doesn't seem to work in docker.

Removed unwanted json files.

Rewrote the README file.