load generation strategies + gradle build
Organized code to allow running a single application with different
strategies for load generation.
Added build.gradle to allow build/clean/run using gradle.

Currently, build.gradle is written such that a fat jar is created.
This is to make sure that all the dependencies are bundled together
with the application.