design: strategy + factory (yaml configs) + builder.
Each different load generator now is a LoadGenerationStrategyI.
Instantiation uses a simple factory that takes the config file
and unmarshals it into the necessary load generator.

Added commandline options for specifying the type of load to be
generated.

Load generator config files are written in YAML and have fixed names
and are stored in the project's root directory.

Code compilation and Jar creation is now performed using Gradle.